+++ packages/avalanche-module/src/handlers/avalanche-send-transaction/utils/get-addresses-by-indices.ts
@@ -7,18 +7,32 @@ export const getAddressesByIndices = async ({
   isChange,
   isTestnet,
   xpubXP,
+  externalXPAddresses,
 }: {
   indices: number[];
   chainAlias: 'X' | 'P';
   isChange: boolean;
   isTestnet: boolean;
   xpubXP?: string;
+  externalXPAddresses?: { index: number; address: string }[];
 }): Promise<string[]> => {
-  if (!xpubXP || (isChange && chainAlias !== 'X')) {
+  if (isChange && chainAlias !== 'X') {
     return [];
   }
 
-  const provider = await getProvider({ isTestnet });
+  if (xpubXP) {
+    const provider = await getProvider({ isTestnet });
 
-  return indices.map((index) => Avalanche.getAddressFromXpub(xpubXP, index, provider, chainAlias, isChange));
+    return indices.map((index) => Avalanche.getAddressFromXpub(xpubXP, index, provider, chainAlias, isChange));
+  }
+
+  if (externalXPAddresses && Array.isArray(externalXPAddresses)) {
+    const accounts = indices
+      .map((index) => externalXPAddresses.find((address) => address.index === index))
+      .filter((acc) => acc !== undefined);
+
+    return accounts.map((account) => account.address);
+  }
+
+  return [];
 };