+++ packages/avalanche-module/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @avalabs/avalanche-module
 
+## 2.2.1
+
+### Patch Changes
+
+- 09b31d8: chore(avalanche): rename xpAddresses to externalXPAddresses
+  - @avalabs/vm-module-types@2.2.1
+
+## 2.2.0
+
+### Minor Changes
+
+- a48d1b1: feat(avalanche-module): pass xp account info
+
+### Patch Changes
+
+- @avalabs/vm-module-types@2.2.0
+
 ## 2.1.0
 
 ### Patch Changes