+++ packages/bitcoin-module/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @avalabs/bitcoin-module
 
+## 2.2.1
+
+### Patch Changes
+
+- @avalabs/vm-module-types@2.2.1
+
+## 2.2.0
+
+### Patch Changes
+
+- @avalabs/vm-module-types@2.2.0
+
 ## 2.1.0
 
 ### Patch Changes