+++ packages/types/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @avalabs/vm-module-types
 
+## 2.2.1
+
+## 2.2.0
+
 ## 2.1.0
 
 ### Minor Changes